docker refactoring

+++ lighthouse/enrollment-server/app.py
@@ -1,6 +1,8 @@
 import subprocess
 import os
 import dotenv
+import requests
+import tarfile
 import flask
 import yaml
 
@@ -9,11 +11,8 @@ import yaml
 # X=1 is the Lighthouse address
 # X=2-254 are the host addresses 
 # X=255 is the broadcast address
-
-TEMPLATE_PATH = "./config-template.yaml"
-CA_PATH = "./ca.crt"
     
-host_id = 1;
+host_id = 1
 app = flask.Flask(__name__)
 
 @app.route('/enroll', methods=['GET'])
@@ -21,12 +20,13 @@ def enroll():
     
     # Authenticate Request
     network_key = flask.request.args.get('network_key')
-    if network_key != NETWORK_KEY:
+    if not network_key or network_key != NETWORK_KEY:
         return "Unauthorized", 401
         
     # Generate Host Configuration
+    global host_id
     config_yaml = generate_nebula_config()
-    config_path = f"./config_{host_id - 2}.yaml",
+    config_path = f"./shared/config_{host_id}.yaml"
     with open(config_path, "w") as config_file:
         yaml.dump(config_yaml, config_file)  
     host_id += 1
@@ -35,26 +35,36 @@ def enroll():
 
 def generate_nebula_config(isLighthouse=False):
     
-    subprocess.run(["nebula-cert", "sign", "-name", f"{host_id}", "-ip", f"192.168.100.{host_id}/24"])
-
+    
     # Load Configuration Template
-    with open(TEMPLATE_PATH, "r") as template:
+    with open("./config-template.yaml", "r") as template:
         config = yaml.safe_load(template)
         
+    # Generate Key and Certificate
+    os.chdir("./shared")
+    subprocess.run(["./nebula-cert", "sign", "-name", f"{host_id}", "-ip", f"192.168.100.{host_id}/24"])
+
+    # TODO: Fix formatting of all 3
     # Add Certificates and Keys to Configuration
-    with open(CA_PATH, "r") as ca_file:
-        config["pki"]["ca"] = f"|\n{ca_file.read()}"
+    with open("./ca.crt", "r") as ca_file:
+        config["pki"]["ca"] = ca_file.read().strip()
         
     with open(f"./{host_id}.crt", "r") as crt_file:
-        config["pki"]["crt"] = f"|\n{crt_file.read()}"
+        config["pki"]["cert"] = crt_file.read().strip()
         
     with open(f"./{host_id}.key", "r") as key_file:
-        config["pki"]["key"] = f"|\n{key_file.read()}"
+        config["pki"]["key"] = key_file.read().strip()
+        
+    # Cleanup
+    os.remove(f"./{host_id}.crt")
+    os.remove(f"./{host_id}.key")
+    os.chdir("..")
         
     # Lighthouse Configuration
     if isLighthouse:
         config["lighthouse"]["am_lighthouse"] = True
-        config["static_host_map"]["\"192.168.100.1\""] = LIGHTHOUSE_PUBLIC_IP
+        # TODO: Fix this formatting
+        config["static_host_map"]["192.168.100.1"] = [LIGHTHOUSE_PUBLIC_IP]
         config["lighthouse"]["hosts"] = ""
     
     return config 
@@ -63,16 +73,39 @@ if __name__ == '__main__':
     
     # Load Environment Variables
     dotenv.load_dotenv()
-    NETWORK_KEY = os.getenv("NETWORK_KEY").split(",")
+    NETWORK_KEY = os.getenv("NETWORK_KEY")
     LIGHTHOUSE_PUBLIC_IP = os.getenv("LIGHTHOUSE_PUBLIC_IP")
+
+    # Download Nebula Certificate Generator if necessary
+    if not os.path.exists("./shared/nebula-cert"):
     
-    # Generate CA Certificate
-    subprocess.run(["nebula-cert", "ca", "-name", "\"Polaris\""])
-    
-    # Generate Lighthouse Configuration
-    config = generate_nebula_config(isLighthouse=True)
-    with open("/etc/nebula/config-lighthouse.yaml", "w") as config_file:
-        yaml.dump(config, config_file)
+        url = "https://github.com/slackhq/nebula/releases/download/v1.9.5/nebula-linux-amd64.tar.gz"
+        response = requests.get(url)
+        with open('./shared/nebula-linux-amd64.tar.gz', 'wb') as file:
+            file.write(response.content)
+
+        with tarfile.open('./shared/nebula-linux-amd64.tar.gz', 'r:gz') as tar:
+            tar.extractall('./shared')
+        
+        os.remove('./shared/nebula-linux-amd64.tar.gz')
+        os.remove('./shared/nebula')
+        print("Nebula Certificate Generator Downloaded.")
+
+    # Generate CA Key if necessary
+    if not os.path.exists("./shared/ca.key"):     
+        os.chdir("./shared")
+        subprocess.run(["./nebula-cert", "ca", "-name", "\"Polaris\""])
+        os.chdir("..")
+        print("CA Key and Certificate Generated.")   
+        
+    # Generate Lighthouse Configuration if necessary
+    if not os.path.exists("./shared/config.yml"):
+        config = generate_nebula_config(isLighthouse=True)
+        with open("./shared/config.yml", "w") as config_file:
+            yaml.dump(config, config_file)
+        host_id += 1
+        print("Lighthouse Configuration Generated.")   
+        
     
     # Start Flask Server
-    app.run(debug=True)
+    app.run(host='0.0.0.0', port=80)
